Fixed bugs with the symlink mechanism.

+++ pintaexec.py
@@ -86,14 +86,16 @@ def run_filterbank(session, item, branch):
         filterbank_in_file = output_file_name(session, item, branch, 'gpt.dat')
     else:
         filterbank_in_file = './' + os.path.basename(item.rawdatafile)
+        
+    if session.input_dir != session.working_dir:
         if os.path.exists(filterbank_in_file):
             print("[INFO] Removing existing symlink {}".format(filterbank_in_file))
             print("[CMD] rm {}".format(filterbank_in_file))
             os.remove(filterbank_in_file)
+        
         print("Creating symlink to the rawdatafile.")
         print("[CMD] ln -s {} {}".format(item.rawdatafile, filterbank_in_file))
         os.symlink(item.rawdatafile, filterbank_in_file)
-        lns = True
     
     fil_file = output_file_name(session, item, branch, 'fil')
     cmd = "filterbank {} -mjd {:0.18f} -rf {} -nch {} -bw {} -ts {} -df {} > {}".format(filterbank_in_file, item.timestamp, item.freq, item.nchan, item.chanwidth, item.tsmpl, item.sideband_code, fil_file)
@@ -107,7 +109,7 @@ def run_filterbank(session, item, branch):
         
         print_exec_time(branch, program, stop_time-start_time)
         
-    if lns:
+    if branch == 'norfix' and session.input_dir != session.working_dir:
         print("Removing symlink to the rawdatafile.")
         print("[CMD] rm {}".format(filterbank_in_file))
         os.remove(filterbank_in_file)
